microservices/patients-ms/requirements.txt
django>=4.2,<7.0
djangorestframework
django-environ
psycopg2-binary
django-cors-headers
requests